Move first unit test to test file
